Application tests added

+++ Application/ShoppingCartItems/Commands/RemoveShoppingCartItemCommand.cs
@@ -1,6 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
 using Application.Interfaces.Persistence;
 using Domain.ShoppingCartItems;
 
@@ -16,6 +14,8 @@ namespace Application.ShoppingCartItems.Commands
         }
         public void Execute(int shopItemId, string sessionId)
         {
+            if(sessionId is null) throw new ArgumentNullException(nameof(sessionId));
+
             _shoppingCartItemRepository.Remove(new ShoppingCartItem()
             {
                 ShopItemId = shopItemId,
